[Mintak Oh] feature/new-vendor-schema: modify schema
[Mintak Oh] feature/new-vendor-schema: modify editorspick api

[Mintak Oh] feature/new-vendor-schema: modify editorspick api

[Mintak Oh] feature/new-vendor-schema: modify vendor schema

[DongSeungLee] feature/new-vendor-schema : remove fetchJoin in MapWholeSalerSisterEntityRepositoryCustomImpl

[Mintak Oh] FashionGo-Feature-QA/1239: fix #FashionGo-Feature-QA/1239

[Mintak Oh] feature/new-vendor-schema: modify new vendor schema

[Mintak Oh] feature/new-vendor-schema: modify get vendor list query

+++ src/main/java/net/fashiongo/webadmin/data/model/sitemgmt/show/ShowPromotionVendorResponse.java
@@ -60,7 +60,7 @@ public class ShowPromotionVendorResponse {
 		return ShowPromotionVendorResponse.builder()
 				.mapId(row.getMapId())
 				.planId(row.getPlanId())
-				.wholesalerId(row.getWholesalerId())
+				.wholesalerId(row.getWholesalerId().intValue())
 				.commissionRate(row.getCommissionRate().doubleValue())
 				.rackCount(row.getRackCount())
 				.fee(row.getFee())
